# contracts/agents-api/core.openapi.yaml
basePath: /
definitions:
...
  api.CoreAPIError:
    properties:
      code:
        type: string
        x-nullable: true
      details:
        description: |-
          Details contains only documented, Core-owned facts: string, finite number,
          boolean, null or string array values. Never include request echoes, secrets
          or native/provider error text. Empty or invalid details are omitted.
        type: object
      message:
        type: string
      param:
        type: string
        x-nullable: true
      type:
        type: string
    required:
      - message
      - type
    type: object
  api.CoreErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.CoreAPIError'
    required:
      - error
    type: object
...
  api.DiagnosticFailure:
    properties:
      code:
        enum:
          - harness_error
          - model_provider_required
          - runtime_unavailable
          - runtime_disconnected
          - runtime_preparation_failed
          - execution_interrupted
          - delivery_unconfirmed
          - input_rejected
          - executor_protocol_error
          - core_storage_failed
          - internal_error
          - environment_connection_timeout
          - environment_unavailable
          - environment_provisioning_failed
          - authentication_error
          - rate_limit_exceeded
          - usage_limit_exceeded
          - server_overloaded
          - server_error
          - invalid_request
          - resource_not_found
          - request_timeout
          - context_length_exceeded
          - cyber_policy
          - connection_failed
        type: string
      failed_at:
        type: string
        x-nullable: true
      params:
        type: object
    type: object
...
  api.ItemDiagnosticTiming:
    properties:
      completed_at:
        type: string
        x-nullable: true
      item_id:
        type: string
      observed_duration_ms:
        type: integer
        x-nullable: true
      started_at:
        type: string
    type: object
...
  api.SessionDiagnosticFailure:
    properties:
      code:
        enum:
          - harness_error
          - model_provider_required
          - runtime_unavailable
          - runtime_disconnected
          - runtime_preparation_failed
          - execution_interrupted
          - delivery_unconfirmed
          - input_rejected
          - executor_protocol_error
          - core_storage_failed
          - internal_error
          - environment_connection_timeout
          - environment_unavailable
          - environment_provisioning_failed
          - authentication_error
          - rate_limit_exceeded
          - usage_limit_exceeded
          - server_overloaded
          - server_error
          - invalid_request
          - resource_not_found
          - request_timeout
          - context_length_exceeded
          - cyber_policy
          - connection_failed
        type: string
      failed_at:
        type: string
        x-nullable: true
      params:
        type: object
      source:
        enum:
          - turn
          - environment
          - environment_input
        type: string
      turn_id:
        type: string
    type: object
  api.SessionDiagnostics:
    properties:
      failure:
        allOf:
          - $ref: '#/definitions/api.SessionDiagnosticFailure'
        x-nullable: true
      object:
        enum:
          - core.session_diagnostics
        type: string
      session_id:
        type: string
      status:
        enum:
          - idle
          - in_progress
          - requires_action
          - failed
        type: string
    type: object
  api.TurnDiagnostics:
    properties:
      failure:
        allOf:
          - $ref: '#/definitions/api.DiagnosticFailure'
        x-nullable: true
      items:
        items:
          $ref: '#/definitions/api.ItemDiagnosticTiming'
        type: array
      items_truncated:
        type: boolean
      object:
        enum:
          - core.turn_diagnostics
        type: string
      session_id:
        type: string
      status:
        type: string
      turn_id:
        type: string
    type: object
...
  v1.Agent:
    properties:
      id:
        type: string
      instructions:
        type: string
        x-nullable: true
      model:
        type: string
      multi_agent:
        $ref: '#/definitions/v1.MultiAgentConfig'
      name:
        type: string
        x-nullable: true
      reasoning:
        $ref: '#/definitions/v1.Reasoning'
      service_tier:
        enum:
          - auto
        type: string
      text:
        $ref: '#/definitions/v1.TextConfig'
      tools:
        items:
          type: object
        type: array
      x_agents_core:
        allOf:
          - $ref: '#/definitions/v1.AgentsCore'
        x-nullable: true
    required:
      - id
      - model
      - multi_agent
      - reasoning
      - service_tier
      - text
      - tools
    type: object
...
  v1.AgentsCore:
    properties:
      harness:
        enum:
          - codex
          - claude_sdk
          - mcode
        type: string
    required:
      - harness
    type: object
...
  v1.EnvironmentNetwork:
    properties:
      access:
        enum:
          - enabled
          - disabled
          - restricted
        type: string
      allowed_domains:
        items:
          type: string
        type: array
    required:
      - access
      - allowed_domains
    type: object
  v1.EnvironmentPackagesResponse:
    properties:
      npm:
        items:
          type: string
        type: array
      python:
        items:
          type: string
        type: array
      system:
        items:
          type: string
        type: array
    required:
      - npm
      - python
      - system
    type: object
...
  v1.InputTokenDetails:
    properties:
      cached_tokens:
        type: integer
    required:
      - cached_tokens
    type: object
...
  v1.MultiAgentConfig:
    properties:
      enabled:
        type: boolean
      max_concurrent_subagents:
        type: integer
        x-nullable: true
    required:
      - enabled
    type: object
...
  v1.OutputTokenDetails:
    properties:
      reasoning_tokens:
        type: integer
    required:
      - reasoning_tokens
    type: object
  v1.Reasoning:
    properties:
      effort:
        type: string
        x-nullable: true
      summary:
        type: string
        x-nullable: true
    type: object
  v1.RequiredAction:
    properties:
      arguments: {}
      call_id:
        type: string
      environment_id:
        type: string
      name:
        type: string
      turn_id:
        type: string
      type:
        enum:
          - function_call
          - environment_connection
        type: string
    required:
      - type
    type: object
...
  v1.Session:
    properties:
      agent:
        $ref: '#/definitions/v1.Agent'
      created_at:
        type: integer
      environment:
        $ref: '#/definitions/v1.SessionEnvironment'
      error:
        type: string
        x-nullable: true
      id:
        type: string
      last_active_at:
        type: integer
      metadata:
        additionalProperties:
          type: string
        type: object
      object:
        enum:
          - agent.session
        type: string
      required_actions:
        items:
          $ref: '#/definitions/v1.RequiredAction'
        type: array
      status:
        enum:
          - idle
          - in_progress
          - requires_action
          - failed
        type: string
      usage:
        allOf:
          - $ref: '#/definitions/v1.TokenUsage'
        x-nullable: true
      vault_ids:
        items:
          type: string
        type: array
    required:
      - agent
      - created_at
      - environment
      - id
      - last_active_at
      - metadata
      - object
      - required_actions
      - status
      - vault_ids
    type: object
...
  v1.SessionDeleted:
    properties:
      deleted:
        type: boolean
      id:
        type: string
      object:
        enum:
          - agent.session.deleted
        type: string
    required:
      - deleted
      - id
      - object
    type: object
  v1.SessionEnvironment:
    properties:
      capability_directories:
        items:
          type: string
        type: array
      files:
        items:
          type: object
        type: array
      id:
        type: string
      network:
        $ref: '#/definitions/v1.EnvironmentNetwork'
      packages:
        $ref: '#/definitions/v1.EnvironmentPackagesResponse'
      plugins:
        items:
          type: object
        type: array
      remote_url:
        type: string
      skills:
        items:
          type: object
        type: array
      type:
        enum:
          - none
          - self_hosted
          - openai_hosted
        type: string
      workspace_directory:
        type: string
    required:
      - type
    type: object
...
  v1.SessionList:
    properties:
      data:
        items:
          $ref: '#/definitions/v1.Session'
        type: array
      first_id:
        type: string
        x-nullable: true
      has_more:
        type: boolean
      last_id:
        type: string
        x-nullable: true
      object:
        enum:
          - list
        type: string
    required:
      - data
      - has_more
      - object
    type: object
...
  v1.TextConfig:
    properties:
      format:
        $ref: '#/definitions/v1.TextFormat'
      verbosity:
        enum:
          - low
          - medium
          - high
        type: string
    required:
      - format
      - verbosity
    type: object
  v1.TextFormat:
    properties:
      schema:
        type: object
      type:
        enum:
          - text
          - json_schema
        type: string
    required:
      - type
    type: object
  v1.TokenUsage:
    properties:
      input_tokens:
        type: integer
      input_tokens_details:
        $ref: '#/definitions/v1.InputTokenDetails'
      output_tokens:
        type: integer
      output_tokens_details:
        $ref: '#/definitions/v1.OutputTokenDetails'
      total_tokens:
        type: integer
    required:
      - input_tokens
      - input_tokens_details
      - output_tokens
      - output_tokens_details
      - total_tokens
    type: object
...
info:
  contact: {}
  description: Deployment and operations routes under /core/v1 for Core Web's server and operator scripts. Every operation requires the Core key; Project API keys and machine credentials are not accepted.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenAgentCore Core API
  version: "1"
paths:
...
  /core/v1/projects/{project_id}/sessions:
    get:
      description: Core key only. Reuses the public resource projection and operation rules; the Project ID selects the target space and does not authenticate.
      parameters:
        - description: Root Agent ID whose Sessions to return
          in: query
          name: agent_id
          type: string
        - description: Last Session ID from the previous page
          in: query
          name: after
          type: string
        - default: 20
          description: Page size; 0 is treated as 1 and values above 100 as 100
          in: query
          minimum: 0
          name: limit
          type: integer
        - default: desc
          description: Creation order; omit for descending, explicit empty values are invalid
          enum:
            - asc
            - desc
          in: query
          name: order
          type: string
        - description: Project ID
          in: path
          name: project_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SessionList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
      security:
        - DeploymentAdminAuth: []
      summary: List execution Sessions in a Project
      tags:
        - Sessions
  /core/v1/projects/{project_id}/sessions/{session_id}:
    delete:
      description: Core key only. Reuses the public resource projection and operation rules; the Project ID selects the target space and does not authenticate.
      parameters:
        - description: Session ID
          in: path
          name: session_id
          required: true
          type: string
        - description: Project ID
          in: path
          name: project_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SessionDeleted'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
      security:
        - DeploymentAdminAuth: []
      summary: Delete an execution Session in a Project
      tags:
        - Sessions
    get:
      description: Core key only. Reuses the public resource projection and operation rules; the Project ID selects the target space and does not authenticate.
      parameters:
        - description: Session ID
          in: path
          name: session_id
          required: true
          type: string
        - description: Project ID
          in: path
          name: project_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Session'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
      security:
        - DeploymentAdminAuth: []
      summary: Retrieve an execution Session in a Project
      tags:
        - Sessions
...
  /core/v1/projects/{project_id}/sessions/{session_id}/diagnostics:
    get:
      description: Core key only. Safe failure categories from one committed snapshot; no native text or historical inference.
      parameters:
        - description: Project ID
          in: path
          name: project_id
          required: true
          type: string
        - description: Session ID
          in: path
          name: session_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SessionDiagnostics'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
      security:
        - DeploymentAdminAuth: []
      summary: Retrieve root Session diagnostics
      tags:
        - Sessions
...
  /core/v1/projects/{project_id}/sessions/{session_id}/turns/{turn_id}/diagnostics:
    get:
      description: Core key only. At most 1000 root Item receipt timings in public Item order. Receipt intervals are not native execution durations.
      parameters:
        - description: Project ID
          in: path
          name: project_id
          required: true
          type: string
        - description: Session ID
          in: path
          name: session_id
          required: true
          type: string
        - description: Root Turn ID
          in: path
          name: turn_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TurnDiagnostics'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
      security:
        - DeploymentAdminAuth: []
      summary: Retrieve root Turn diagnostics
      tags:
        - Turns
...
schemes:
  - http
  - https
securityDefinitions:
  DeploymentAdminAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"